cleanup

+++ plugins/robots/common/twoDModel/include/twoDModel/engine/model/constants.h
@@ -28,7 +28,8 @@ const QSize displaySize(200, 300);
 const qreal beepWavesSize = 120;
 const qreal gyroscopeConstant = -800;
 const qreal accelerometerConstant = 30000;
-const qreal g = 4065;
+
+const int g = 4065;
 
 const qreal robotWheelDiameterInPx = 16;
 const qreal robotWheelDiameterInCm = 5.6;
